refactor the folder
