+++ go/base/pluginmanager/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["interface.go"],
-    importpath = "github.com/michelangelo-ai/michelangelo/go/components/deployment/utils/pluginmanager",
+    importpath = "github.com/michelangelo-ai/michelangelo/go/base/pluginmanager",
     visibility = ["//visibility:public"],
     deps = [
         "//proto/api/v2:go_default_library",
