+++ go/base/pluginmanager/interface.go
