+++ go/base/revision/interface.go
