+++ go/components/deployment/BUILD.bazel
@@ -18,10 +18,10 @@ go_library(
         "//go/base/conditions/engine:go_default_library",
         "//go/base/conditions/interfaces:go_default_library",
         "//go/base/env:go_default_library",
+        "//go/base/pluginmanager:go_default_library",
+        "//go/base/revision:go_default_library",
         "//go/components/deployment/plugins:go_default_library",
         "//go/components/deployment/plugins/noop:go_default_library",
-        "//go/components/deployment/utils/pluginmanager:go_default_library",
-        "//go/components/deployment/utils/revision:go_default_library",
         "//go/logging:go_default_library",
         "//proto/api:go_default_library",
         "//proto/api/v2:go_default_library",
