+++ go/components/deployment/controller.go
@@ -30,10 +30,10 @@ import (
 	"github.com/michelangelo-ai/michelangelo/go/api/utils"
 	defaultengine "github.com/michelangelo-ai/michelangelo/go/base/conditions/engine"
 	conditionInterfaces "github.com/michelangelo-ai/michelangelo/go/base/conditions/interfaces"
+	"github.com/michelangelo-ai/michelangelo/go/base/pluginmanager"
+	"github.com/michelangelo-ai/michelangelo/go/base/revision"
 	"github.com/michelangelo-ai/michelangelo/go/components/deployment/plugins"
 	"github.com/michelangelo-ai/michelangelo/go/components/deployment/plugins/noop"
-	"github.com/michelangelo-ai/michelangelo/go/components/deployment/utils/pluginmanager"
-	"github.com/michelangelo-ai/michelangelo/go/components/deployment/utils/revision"
 	"github.com/michelangelo-ai/michelangelo/go/logging"
 	protoapi "github.com/michelangelo-ai/michelangelo/proto/api"
 	v2pb "github.com/michelangelo-ai/michelangelo/proto/api/v2"
